rocm-smi: Use better file to test for PowerPlay availability
The pp_table file might not be present on all machines, but certain
features like Performance Level can still be used. Check for the
bare-minimum for usability, the power_dpm_force_performance_level. All
features are checked on usage, so this is a safer bet for determining
minimum PowerPlay usage.

Change-Id: I56e8f21eccabc90133bcd78d645d3bd5c0223421