SMI: Cleanup setting Power Profile, fix setprofile test
Clean this up so that we always verify the data first, handling both
string and list values. And ensure that Performance Level is set to
auto, otherwise setting the profile will fail. Lastly, clarify some
things in comments and the README to help explain things
Also fix the --setprofile test, as we need to wait for the clocks to
ramp to the desired setting

Change-Id: I8e3234e694aafb2fc870dc27598fde25a184a428