SMI: Get Power using DRM enumeration instead of PCI device
The /sys/kernel/debug/dri/#/name debugfs file isn't always being
populated, so relying on it to correlate to the PCI device is not
viable. Instead, use the fact that debugfs and sysfs both enumerate the
DRM device via device minor to correlate the devices. Also ensure that
we re-launch as Sudo for reading power, since debugfs needs sudo to read
it

Change-Id: I6efc35854b775672d6242735deb225c96d49c305