Fix python2/3 compatibility issue
Python3 uses input, while python2 uses raw_input. If raw_input exists,
then we know that it's python2, since python3 renamed it to input, so
call raw_input instead of input. This won't pass pylint, but it's a
better workaround than trying to force an install of a package via pip
that can't be added to the DEB/RPM control files

Change-Id: I443af64d760199324887c415d28f8d0b3bd09f28