Add ability to change package power cap in OD mode
We can read and change the maximum power available to a GPU
package through amdgpu sysfs entries. This limit is enforced by
the hardware, so setting it lower than the default can be useful
to reduce a GPU's power draw. This patch allows users of rocm-smi
to both read the current maximum power and to change it (or reset
it to the default).

In corrent ROCm releases, the range of available values can only
be lower than the default. However, there is currently an amdgpu
patch in flight to use this same interface to allow users to raise
the GPU's maximum package power above the GPU's default. As such,
this feature should use the overdrive warning/verification flag,
since setting the power cap above the default is not covered under
AMD warranty.

Change-Id: I888c7421dc0d177feef4e62be6528cd8e728e9b6