Properly allow setting more than one DPM mask
Previously, by concatenating all of the values for SetClocks
together without spaces between them, we would try to combine
multiple requested DPM states into a single string of integers.
For example, trying to set "--setsclk 0 2 4 6" should "turn on"
SCLK DPM states 0, 2, 4, and 6. This would try to set 0246 which
resulted in wrong masks being put into the driver.

We should still check that all the values being sent in are
integers, so this patch first makes one fully-concatenated
variable and ensures that it is still an integers. However, it
actually passes in to sysfs a value that is space separated.

Change-Id: I02b17f716efd70f8f10a1af116282b62e21d4f01