Don't include non-AMD devices by default
Only list/control AMD devices unless the --alldevices flag is passed in,
this will avoid printing a bunch of unsupported fields, since it's not
up to our kernel to support them. And avoid a bunch of error message
regarding being unable to set values on devices that aren't AMD GPUs

Some things may be usable, depending on the driver code in the upstream
kernel (like temperature or fan speed on certain GPUs), but for the most
part, it's stuff that the SMI can't control, which just wastes screen space,
and most users don't care about trying to manage AMD and non-AMD devices in
the same tool. The ones that do can use --alldevices

Change-Id: I5be9458b7194969a8ce16401c9dee47235fb5dc1