Clean up concise output and header/footer
1) Remove PCIe clocks from Concise info. Almost no one will change this
since it's either PCIe gen1 x1 or whatever the max available bandwidth
is
2) When the SMI was in its infancy, using the string processors made
things look pretty with alignment and such. As the SMI has expanded,
maintaining it has become unmanageable. Dynamically check the
length of column headers and column values so that we can add values
more easily.
3) Start trying to keep an 80-char max for output, starting with the
header/footer
Change-Id: Ib04f5fac992f6ecd732c78066485c74c2d063395