Add logging to printLog function
Just in case someone is logging to a file, add logging to printLog to
ensure that all information is logged together. Also remove the check
for logging level for showConcise. Since showConcise and showConciseHW
don't use printLog, using the logging flag is unnecessary, and also adds
too many potential cases for sys.argv length checking. This makes it
simpler and cleaner and easier

Change-Id: I7fe8447de04cb8864d15ad1ef572aaff0a6ce2a1