Show Firmware Versions
Show firmware versions stored in sysfs files (similar to showRasInfo).
This can now be done with ./rocm-smi --showfwinfo <arguments>
where arguments can be 'all' or blank to display all firmwares,
or it can be a list of firmwares seperated by spaces, i.e:
./rocm-smi --showfwinfo vce uvd mc ...etc
The firmware versions are stored as hex in sysfs, but displayed in dec.
The only exceptions currently are the smc fw version, which is encoded
and ta_ras/ta_xgmi/uvd/vce/vcn, which are displayed as hex.

Change-Id: Ifaacf846952e75d5e6108ed4e04131b81b256caf
Signed-off-by: Ori Messinger <ori.messinger@amd.com>