Add support for JSON output
Add a --json flag to format the output of the SMI into JSON format. This
will output whatever flags are requested (so rocm-smi --json -t -P will
print JSON format of temperature and power for each node)

Note that this won't work for certain outputs that are either specially
formatted or are multi-line. An error will alert the user to this.
The -a flag has been modified to exclude these multi-line outputs to
allow for "rocm-smi -a --json" to work as expected
Change-Id: I94146c72785f31770d5418ab4f9dc724d1f0ffb3