Fix device info for mGPU & Vendor check
The file read was indented incorrectly, so only the last device
in the list was processed. The indentation has now been fixed.

Additionally, vendor ID is now checked to ensure the PCI device
being checked is an AMD device and not some other PCI device.
(This is done to prevent issues when --showproductname is called
with the --alldevices flag)

Change-Id: Ic5bec3edadbbd5247fbf58b9cb730016bf9e337e
Signed-off-by: Ori Messinger <ori.messinger@amd.com>