FIx clock printout if DPM is disabled
If DPM is disabled for a block, it will print a single line for the
current speed, and that's it. fclk does not have the * denoting that it
is the current speed, so if we have a single line in the pp_dpm file,
just print it since that's the only one. If there's a warning in the
sysfs file saying "DPM disabled", communicate that in the debug log

Change-Id: Ic18fbd682c5abc5cded55680379dd65af5875b64