Clean up JSON printing
We can still print viable JSON output, even if a bad flag was passed in.
Debug-log the error but continue to print the rest of the data out.
This also aids in consistency, since we won't arbitrarily exit the SMI
depending on which arguments are parsed first.

Also ensure that we don't print an empty set (if 1 bad flag was passed
in, it would print out {card0: {}, card1: {}} , which is useless)

Lastly, clean up a wayward tab from the previous commit

Change-Id: I34e8ce1e5f8bf37ce9a654c4157818cad03b2016